+++ src/form-elements/FormElementFiles.tsx
@@ -45,6 +45,7 @@ function FormElementFiles({
     element,
     onChange,
     setIsDirty,
+    element.maxFileSize,
   )
 
   const inputRef = React.useRef<HTMLInputElement>(null)