+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - validation icon
 - `maxFileSize` validation for `files` element
 
+### Changed
+
+- CSS class for quill content from `ql-editor` to `ob-quill-content`
+
 ## [8.6.0] - 2025-07-01
 
 ### Added