+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `maxFileSize` validation for `files` element
 - `dateTimeCompleted` to form store table
 - Dimensions to drawing tool for `arcGISWebMap` element
+- handling for ArcGIS web map layer visibility
 
 ### Changed