+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Changed
+
+- `@oneblink/apps`
+
 ## [8.6.0] - 2025-07-01
 
 ### Added